.github/Databricks:kafka:ELk/databricks-gcp/requirements.txt
# Secret Manager Requirements
# Python dependencies for secret monitoring automation

google-cloud-secret-manager>=2.16.0
google-cloud-monitoring>=3.15.0
google-cloud-pubsub>=2.18.4
google-cloud-functions>=1.16.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0

# Data processing and utilities
requests>=2.31.0
jq>=1.6.0
python-dateutil>=2.8.2
pytz>=2023.3

# Email and notification support
smtplib-ssl>=1.0.0
email-validator>=2.0.0

# CLI and formatting
click>=8.1.0
tabulate>=0.9.0
colorama>=0.4.6

# Testing and development
pytest>=7.4.0
pytest-mock>=3.11.0
responses>=0.23.0